Rename docdown.template_adapters.pystache_adapter
* Renamed docdown.template_adapters.pystache_adapter to
  docdown.template_adapters.pystache for consistency, shorter name,
  and removal of redundant info in names
