+++ docdown/template_adapters/__init__.py
@@ -2,6 +2,6 @@
 
 from __future__ import absolute_import, unicode_literals, print_function
 
-from .pystache_adapter import PystacheAdapter
+from .pystache import PystacheAdapter
 from .string_format import StringFormatAdapter
 from .template_string import TemplateStringAdapter
